fxbot/api/requirements.txt
fastapi==0.115.5
sqlalchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
uvicorn==0.32.1
httpx==0.28.1